+++ NEWS.md
@@ -1,3 +1,8 @@
+# emphatic 0.1.2  2020-11-11
+
+* Added support for tibbles (by internally converting to data.frames 
+during output)
+
 # emphatic 0.1.1  2020-11-01
 
 * Added bespoke html rendering. Removed dependency on `fansi`